requirements.txt
playwright>=1.40.0
requests>=2.31.0
typer[all]>=0.9.0
rich>=13.7.0
pytesseract>=0.3.10
pillow>=10.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
aiofiles>=23.2.0
asyncio-throttle>=1.0.2 